% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAscanImport-checks.R
\docType{methods}
\name{checktRNAscanGRanges}
\alias{checktRNAscanGRanges}
\alias{checktRNAscanGRanges}
\alias{checktRNAscanGRanges,GRanges-method}
\title{tRNAscan compatibility check}
\usage{
checktRNAscanGRanges(gr)

\S4method{checktRNAscanGRanges}{GRanges}(gr)
}
\arguments{
\item{gr}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{checktRNAscanGRanges} checks whether a GRanges object contains the 
information expected for a tRNAscan result.
}
\examples{
file <- system.file("extdata", 
                    file = "sacCer3-tRNAs.ss.sort", 
                    package = "tRNAscanImport")
gr <- tRNAscanImport::import.tRNAscanAsGRanges(file)
checktRNAscanGRanges(gr)
}
