% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-dotbracket.R
\docType{methods}
\name{gettRNABasePairing}
\alias{gettRNABasePairing}
\alias{getBasePairing}
\alias{gettRNABasePairing}
\alias{gettRNABasePairing,GRanges-method}
\alias{getBasePairing}
\title{Accessing Dot Bracket annotation}
\usage{
gettRNABasePairing(gr)

\S4method{gettRNABasePairing}{GRanges}(gr)

getBasePairing(dotBracket)
}
\arguments{
\item{gr}{a GRanges object created by \code{import.tRNAscanAsGRanges} or 
GRanges with equivalent information. The \code{tRNA_str} column will be used
for input into \code{getBasePairing}.}

\item{dotBracket}{character vectors describing a nucleotide sequence
structure in the dot bracket annotations. Valid characters are:
\code{.(\\{[><]\\})}}
}
\value{
The result is a data.frame with three columns: pos, forward, reverse. If a
position is unpaired, forward and reverse will be \code{0}, otherwise it 
will match the base paired positions.
}
\description{
\code{getBasePairing} converts a dot bracket annotation into a
\code{data.frame}. Base pairing is indicated by corresponding numbers 
in the forward and reverse columns.
}
