% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/tRNAScan2GRanges-plot.R
\docType{methods}
\name{gettRNAscanSummary}
\alias{gettRNAscanSummary}
\alias{plottRNAscanSummary}
\alias{gettRNAscanSummary}
\alias{plottRNAscanSummary}
\alias{plottRNAscanSummary,GRanges-method}
\title{tRNAscan2 Summary}
\usage{
gettRNAscanSummary(gr)

plottRNAscanSummary(gr)

\S4method{gettRNAscanSummary}{GRanges}(gr)

\S4method{plottRNAscanSummary}{GRanges}(gr)
}
\arguments{
\item{gr}{a GRanges object created by the tRNAscan2GRanges or GRanges with
equivalent information.}
}
\value{
\code{gettRNAscanSummary()}: DataFrame
\code{gettRNAscanSummary()}: ggplot
}
\description{
\code{gettRNAscanSummary()}: creates an DataFrame with aggregates of 
information
\code{plottRNAscanSummary()}: If ggplot2 is installed a plot of the data
is generated
}
\examples{
gr <- tRNAscan2GRanges(system.file("extdata", 
                             file = "tRNAscan.sort", 
                             package = "tRNAscan2GRanges"))
gettRNAscanSummary(gr)
plottRNAscanSummary(gr)
}
