% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAscanImport-methods.R
\docType{methods}
\name{gettRNAstructureSeq}
\alias{gettRNAstructureSeq}
\alias{gettRNAstructureSeq}
\alias{gettRNAstructureSeq,GRanges-method}
\title{Subsetting tRNA sequences}
\usage{
gettRNAstructureSeq(gr, structure = "", padSequences = TRUE,
  pad5prime = TRUE)

\S4method{gettRNAstructureSeq}{GRanges}(gr, structure = "",
  padSequences = TRUE, pad5prime = TRUE)
}
\arguments{
\item{gr}{a GRanges object created by \code{import.tRNAscanAsGRanges} or 
GRanges with equivalent information.}

\item{structure}{optional parameter for returning just partial sequences.
The following values are accepted: 
\code{anticodonloop}, \code{Dloop}, \code{Tloop}, \code{acceptorStem}, 
\code{anticodonStem}, \code{DStem}, \code{TStem}, \code{variableLoop}, 
\code{discriminator}. (default = \code{structure = ""})}

\item{padSequences}{optional parameter whether sequences of the same type 
should be returned with the same length. For stems missing positions will be
filled up in the middle, for loops at the ends. 
(default = \code{padSequences = TRUE})}

\item{pad5prime}{If missing sequence length is an uneven number, 
should the 5'-end be padded by one more? (Only applies to loop sequences,
default = \code{padLeft = TRUE})}
}
\value{

}
\description{
\code{gettRNAstructureSeq} returns split or partial tRNA sequences based on
the structure information of tRNAscan.
}
\examples{

}
