% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-checks.R
\docType{methods}
\name{istRNAGRanges}
\alias{istRNAGRanges}
\alias{istRNAGRanges}
\alias{istRNAGRanges,GRanges-method}
\title{tRNA compatibility check}
\usage{
istRNAGRanges(gr)

\S4method{istRNAGRanges}{GRanges}(gr)
}
\arguments{
\item{gr}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{istRNAGRanges} checks whether a GRanges object contains the
information expected for a tRNA result.
}
\examples{
file <- system.file("extdata", 
                    file = "sacCer3-tRNAs.ss.sort", 
                    package = "tRNAscanImport")
gr <- tRNAscanImport::import.tRNAscanAsGRanges(file)
istRNAGRanges(gr)
}
