% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAscanImport-checks.R
\name{istRNAscanGRanges}
\alias{istRNAscanGRanges}
\alias{istRNAscanGRanges,GRanges-method}
\title{tRNAscan compatibility check}
\usage{
istRNAscanGRanges(gr)

\S4method{istRNAscanGRanges}{GRanges}(gr)
}
\arguments{
\item{gr}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{istRNAscanGRanges} checks whether a GRanges object contains the 
information expected for a tRNAscan result.
}
\examples{
file <- system.file("extdata", 
                    file = "yeast.tRNAscan", 
                    package = "tRNAscanImport")
gr <- tRNAscanImport::import.tRNAscanAsGRanges(file)
istRNAscanGRanges(gr)
}
