% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAscanImport-subset.R
\docType{methods}
\name{hasTStem}
\alias{hasTStem}
\alias{hasDStem}
\alias{hasAcceptorStem}
\alias{hasAnticodonStem}
\alias{hasTloop}
\alias{hasDloop}
\alias{hasAnticodonLoop}
\alias{hasVariableLoop}
\alias{tRNA-subset}
\alias{hasTStem,GRanges-method}
\alias{hasDStem,GRanges-method}
\alias{hasAcceptorStem,GRanges-method}
\alias{hasAnticodonStem,GRanges-method}
\title{Subsetting tRNA}
\usage{
hasTStem(gr, unpaired = NA, length = NA, mismatches = NA)

hasDStem(gr, unpaired = NA, length = NA, mismatches = NA)

hasAcceptorStem(gr, unpaired = NA, length = NA, mismatches = NA)

hasAnticodonStem(gr, unpaired = NA, length = NA, mismatches = NA)

hasTloop(gr, length = NA)

hasDloop(gr, length = NA)

hasAnticodonLoop(gr, length = NA)

hasVariableLoop(gr, paired = NA, length = NA, mismatches = NA)

\S4method{hasTStem}{GRanges}(gr, unpaired, length)

\S4method{hasDStem}{GRanges}(gr, unpaired, length)

\S4method{hasAcceptorStem}{GRanges}(gr, unpaired, length)

\S4method{hasAnticodonStem}{GRanges}(gr, unpaired, length)
}
\arguments{
\item{gr}{a GRanges object from a tRNAscan import or with equivalent
information}
}
\value{
logical vector
}
\description{
title
}
