% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAscan2gffAndFa.R
\name{tRNAScan2gffAndFa}
\alias{tRNAScan2gffAndFa}
\title{tRNAScan2gffAndFa

\code{tRNAScan2gffAndFa} will convert a tRNAscan-SE output file into a 
GRanges object.}
\usage{
tRNAScan2gffAndFa(file, cluster = TRUE)
}
\arguments{
\item{file}{tRNAscan-SE input file}
}
\value{
a GRanges object
}
\examples{
tRNAScan2gffAndFa(system.file("tRNAscan.txt",package="tRNAScan2gffAndFa"))
}
