% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAScan2GRanges.R
\name{tRNAscan2GRanges}
\alias{tRNAscan2GRanges}
\title{tRNAscan2GRanges}
\usage{
tRNAscan2GRanges(file, trim_intron = TRUE)
}
\arguments{
\item{file}{tRNAscan-SE input file}

\item{trim_intron}{optional logical: remove intron sequences}
}
\value{
a GRanges object
}
\description{
\code{tRNAScan2GRanges} will convert a tRNAscan-SE output file into a 
GRanges object. tRNAscan-SE 1.3.1 output is expected. Intron sequences are 
removed by default, but can also returned untouched.
}
\examples{
tRNAscan2GRanges(system.file("extdata", 
                             file = "tRNAscan.sort", 
                             package = "tRNAscan2GRanges"))
}
