% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAScan2GRanges-plot.R
\docType{methods}
\name{gettRNAscanSummary}
\alias{gettRNAscanSummary}
\alias{plottRNAscanSummary}
\alias{plottRNAscanSummary}
\alias{gettRNAscanSummary,GRanges-method}
\alias{plottRNAscanSummary,GRangesList-method}
\title{tRNAscan2GRanges Summary}
\usage{
gettRNAscanSummary(gr)

plottRNAscanSummary(grl)

\S4method{gettRNAscanSummary}{GRanges}(gr)

\S4method{plottRNAscanSummary}{GRangesList}(grl)
}
\arguments{
\item{gr}{a GRanges object created by the tRNAscan2GRanges or GRanges with
equivalent information.}

\item{grl}{a GRangesList object created with GRanges created by the 
tRNAscan2GRanges or GRanges with equivalent information.}
}
\value{
\code{gettRNAscanSummary()}: DataFrame
\code{plottRNAscanSummary()}: list of ggplots per column of data
}
\description{
\code{gettRNAscanSummary()}: creates an DataFrame with aggregates of 
information
\code{plottRNAscanSummary()}: If ggplot2 is installed a plot of the data
is generated
}
\examples{
gr <- tRNAscan2GRanges(system.file("extdata", 
                             file = "sacCer3-tRNAs.ss.sort", 
                             package = "tRNAscan2GRanges"))
gettRNAscanSummary(gr)
plots <- plottRNAscanSummary(GenomicRanges::GRangesList(Sce = gr))
}
