% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAscanImport-plot.R
\docType{methods}
\name{gettRNAscanSummary}
\alias{gettRNAscanSummary}
\alias{plottRNAscan}
\alias{gettRNAscanPlots}
\alias{plottRNAscan}
\alias{gettRNAscanPlots}
\alias{gettRNAscanSummary,GRanges-method}
\alias{plottRNAscan,GRangesList-method}
\alias{gettRNAscanPlots,GRangesList-method}
\title{tRNAscanImport summary functions}
\usage{
gettRNAscanSummary(gr)

plottRNAscan(grl)

gettRNAscanPlots(grl)

\S4method{gettRNAscanSummary}{GRanges}(gr)

\S4method{plottRNAscan}{GRangesList}(grl)

\S4method{gettRNAscanPlots}{GRangesList}(grl)
}
\arguments{
\item{gr}{a GRanges object created by \code{import.tRNAscanAsGRanges} or 
GRanges with equivalent information.}

\item{grl}{a GRangesList object created with GRanges created by 
\code{import.tRNAscanAsGRanges} or GRanges with equivalent information.}
}
\value{
\code{gettRNAscanSummary()}: DataFrame

\code{gettRNAscanPlots()}: list of ggplots per column of data

\code{plottRNAscan()}:  directly plots the output of gettRNAscanPlots
}
\description{
\code{gettRNAscanSummary()}: creates an DataFrame with aggregated information
from the tRNAscan information.

\code{gettRNAscanPlots()}: If ggplot2 is installed a plot of the data
is generated and returned.

\code{plottRNAscan()}: If ggplot2 is installed a plot of the data
is generated and ploted directly.
}
\examples{
library(GenomicRanges, quietly = TRUE)
sce <- import.tRNAscanAsGRanges(system.file("extdata",
                               file = "sacCer3-tRNAs.ss.sort",
                               package = "tRNAscanImport"))
eco <- import.tRNAscanAsGRanges(system.file("extdata",
                        file = "eschColi_K_12_MG1655-tRNAs.ss.sort",
                        package = "tRNAscan2GRanges"))
gettRNAscanSummary(sce)
plots <- gettRNAscanPlots(GRangesList(Sce = sce,
                                      Eco = eco))
}
