% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAscanImport-methods.R
\docType{methods}
\name{gettRNAstructureGRanges}
\alias{gettRNAstructureGRanges}
\alias{gettRNAstructureSeqs}
\alias{gettRNAstructureSeqs}
\alias{gettRNAstructureGRanges}
\alias{gettRNAstructureGRanges,GRanges-method}
\alias{gettRNAstructureSeqs,GRanges-method}
\title{Subsetting tRNA sequences}
\usage{
gettRNAstructureGRanges(gr, structure = "")

gettRNAstructureSeqs(gr, structure = "", joinCompletely = FALSE,
  joinFeatures = TRUE, padSequences = TRUE, padCenter = TRUE,
  pad5prime = FALSE)

\S4method{gettRNAstructureGRanges}{GRanges}(gr, structure = "")

\S4method{gettRNAstructureSeqs}{GRanges}(gr, structure = "",
  joinCompletely = FALSE, joinFeatures = TRUE, padSequences = TRUE,
  padCenter = TRUE, pad5prime = FALSE)
}
\arguments{
\item{gr}{a GRanges object created by \code{import.tRNAscanAsGRanges} or 
GRanges with equivalent information.}

\item{structure}{optional parameter for returning just partial sequences.
The following values are accepted: 
\code{anticodonloop}, \code{Dloop}, \code{Tloop}, \code{acceptorStem}, 
\code{anticodonStem}, \code{DStem}, \code{TStem}, \code{variableLoop}, 
\code{discriminator}. (default = \code{structure = ""})}

\item{joinCompletely}{Should the sequence parts, which are to be returned, be
joined into one sequence? (default = \code{joinCompletely = FALSE}))
Setting this to TRUE excludes \code{joinFeatures} be set to TRUE as well. In
addition, \code{joinCompletely = TRUE} uses automatically all sequence
structures.}

\item{joinFeatures}{Should the sequence parts, which are to be returned and
are from the same structure type, be joined into one sequence?
(default = \code{joinCompletely = TRUE})) Setting this to TRUE excludes 
\code{joinCompletely} be set to TRUE as well.}

\item{padSequences}{parameter whether sequences of the same type 
should be returned with the same length. For stems missing positions will be
filled up in the middle, for loops at the ends. 
(default = \code{padSequences = TRUE})}

\item{padCenter}{If missing sequence length is an uneven number, 
should the 5'-end be padded by one more? (Only applies to loop sequences,
default = \code{padCenter = TRUE})}

\item{pad5prime}{If missing sequence length is an uneven number, 
should the 5'-end be padded by one more? (Only applies to loop sequences,
default = \code{pad5prime = FALSE})}
}
\value{
a list of \code{GRanges} or \code{DNAStringSet} objects. In case
joinCompletly is set to TRUE a single \code{DNAStringSet} is returned.
}
\description{
\code{gettRNAstructureSeq} returns split or partial tRNA sequences based on
the structure information of tRNAscan. Variances in length of certain 
structure features can be padded.

Padding works upon certain rules, which can be controlled via 
\code{padCenter} and \code{pad5prime}. Both parameters only affect 
padding of loops. With \code{padCenter = TRUE} length is adjusted by adding
"-" characters in the middle, whereas \code{padCenter = FALSE} adds them at
each end. If \code{pad5prime = TRUE} more padding characters are added on the
5'-end if there are uneven numbers of missing positions.
}
\examples{
gr <- import.tRNAscanAsGRanges(system.file("extdata", 
                               file = "sacCer3-tRNAs.ss.sort", 
                               package = "tRNAscanImport"))
gettRNAstructureGRanges(gr, structure = "anticodonloop")
gettRNAstructureSeqs(gr, structure = "anticodonloop")
}
