% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-subset.R
\docType{methods}
\name{hasTStem}
\alias{hasTStem}
\alias{hasDStem}
\alias{hasAcceptorStem}
\alias{hasAnticodonStem}
\alias{hasTloop}
\alias{hasDloop}
\alias{hasAnticodonLoop}
\alias{hasVariableLoop}
\alias{tRNA-subset}
\alias{hasTStem,GRanges-method}
\alias{hasDStem,GRanges-method}
\alias{hasAcceptorStem,GRanges-method}
\alias{hasAnticodonStem,GRanges-method}
\alias{hasTloop,GRanges-method}
\alias{hasDloop,GRanges-method}
\alias{hasAnticodonLoop,GRanges-method}
\alias{hasVariableLoop,GRanges-method}
\title{Subsetting tRNA}
\usage{
hasTStem(gr, length = NA, unpaired = NA, mismatches = NA, bulged = NA)

hasDStem(gr, length = NA, unpaired = NA, mismatches = NA, bulged = NA)

hasAcceptorStem(gr, length = NA, unpaired = NA, mismatches = NA,
  bulged = NA)

hasAnticodonStem(gr, length = NA, unpaired = NA, mismatches = NA,
  bulged = NA)

hasTloop(gr, length = NA)

hasDloop(gr, length = NA)

hasAnticodonLoop(gr, length = NA)

hasVariableLoop(gr, length = NA, paired = NA, mismatches = NA,
  bulged = NA)

\S4method{hasTStem}{GRanges}(gr, length = NA, unpaired = NA,
  mismatches = NA, bulged = NA)

\S4method{hasDStem}{GRanges}(gr, length = NA, unpaired = NA,
  mismatches = NA, bulged = NA)

\S4method{hasAcceptorStem}{GRanges}(gr, length = NA, unpaired = NA,
  mismatches = NA, bulged = NA)

\S4method{hasAnticodonStem}{GRanges}(gr, length = NA, unpaired = NA,
  mismatches = NA, bulged = NA)

\S4method{hasTloop}{GRanges}(gr, length = NA)

\S4method{hasDloop}{GRanges}(gr, length = NA)

\S4method{hasAnticodonLoop}{GRanges}(gr, length = NA)

\S4method{hasVariableLoop}{GRanges}(gr, length = NA, paired = NA,
  mismatches = NA, bulged = NA)
}
\arguments{
\item{gr}{a GRanges object from a tRNAscan import or with equivalent
information}

\item{length}{the length as integer}

\item{unpaired}{logical: has unpaired nucleotides}

\item{mismatches}{logical: has mismatched nucleotides}

\item{bulged}{logical: has mismatched nucleotides of different length
creating a bulge}

\item{paired}{logical: has paired nucleotides (only used for loops)}
}
\value{
a logical vector of the length or input GRanges object
}
\description{
title
}
\examples{
gr <- import.tRNAscanAsGRanges(system.file("extdata", 
                               file = "sacCer3-tRNAs.ss.sort", 
                               package = "tRNAscanImport"))
hasTStem(gr, length = 5, mismatches = TRUE)
gr[hasTStem(gr, length = 5, mismatches = TRUE)]
}
