% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAScan2GRanges.R
\name{tRNAScan2GRanges}
\alias{tRNAScan2GRanges}
\title{tRNAScan2GRanges

\code{tRNAScan2GRanges} will convert a tRNAscan-SE output file into a 
GRanges object.}
\usage{
tRNAScan2GRanges(file, trim_intron = TRUE)
}
\arguments{
\item{file}{tRNAscan-SE input file}

\item{trim_intron}{optional logical: remove intron sequences}
}
\value{
a GRanges object
}
\examples{
tRNAScan2GRanges(system.file("tRNAscan.txt",package="tRNAScan2GRanges"))
}
