% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAScan2GRanges.R
\name{tRNAscan2GRanges}
\alias{tRNAscan2GRanges}
\alias{tRNAscan2GFF}
\title{tRNAscan2GRanges tRNAscan2GFF}
\usage{
tRNAscan2GRanges(file, trim_intron = TRUE)

tRNAscan2GFF(file, trim_intron = TRUE)
}
\arguments{
\item{file}{tRNAscan-SE input file}

\item{trim_intron}{optional logical: remove intron sequences}
}
\value{
a GRanges object
}
\description{
\code{tRNAScan2GRanges} will convert a tRNAscan-SE output file into a 
GRanges object. tRNAscan-SE 1.3.1 output is expected. Intron sequences are 
removed by default, but can also returned untouched.
\code{tRNAScan2GFF} formats the output of \code{tRNAScan2GRanges} to be GFF3
compliant
}
\examples{
tRNAscan2GRanges(system.file("extdata", 
                             file = "sacCer3-tRNAs.ss.sort", 
                             package = "tRNAscan2GRanges"))
                             
tRNAscan2GFF(system.file("extdata", 
                         file = "sacCer3-tRNAs.ss.sort", 
                         package = "tRNAscan2GRanges"))
}
\references{
Chan, Patricia P., and Todd M. Lowe. 2016. “GtRNAdb 2.0: An Expanded Database
of Transfer Rna Genes Identified in Complete and Draft Genomes.” Nucleic
Acids Research 44 (D1): D184–9. doi:10.1093/nar/gkv1309.

Lowe, T. M., and S. R. Eddy. 1997. “TRNAscan-Se: A Program for Improved
Detection of Transfer Rna Genes in Genomic Sequence.” Nucleic Acids Research
25 (5): 955–64.
}
